Debug logging configuration (load at runtime)
